museDemo/requirements.txt
matplotlib
numpy
scikit-learn
scipy
pylsl
muselsl
python-osc
paramiko
google-generativeai